Update raster-methods.R (new function: interpolate.maps)
New function: interpolate.maps

Given a limited number of monthly maps, interpolate map values for the missing months of the year (or any other cyclical map series) using a sine function. Includes a 'supplement' option - making a first rough fit, picking one month from the results and including that as a fixed value in a second round, which prevents excessive amplitudes in the fitted values.