Update ships_methods.R (fix: remove routes that have no movement)
I found that interpolate.ais.data successfully removes ships that have too fews steps to be interpolated, but has no check for ships that have no movement. These are written to the JSON as just one long 'pause' that spans the duration of the file, with no impact on simulation (no noise) except to clutter up the visuals, and should be removed.
Fix implementation is not super elegant because I couldn't readily figure out how to skip them in interpolate.asi.data, so I'm cutting them out of the final object in ais.to.Deponsships. Can't vectorize over sub-entries in a list, thus looping, but time requirement is negligible.